requirements.txt
pandas
flask
azure-storage-blob
requests
